% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{add_boost_tree_catboost}
\alias{add_boost_tree_catboost}
\title{Wrapper to add \code{catboost} engine to the parsnip \code{boost_tree} model
specification}
\usage{
add_boost_tree_catboost()
}
\description{
Wrapper to add \code{catboost} engine to the parsnip \code{boost_tree} model
specification
}
