% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{add_boost_tree_lightgbm}
\alias{add_boost_tree_lightgbm}
\title{Wrapper to add \code{lightgbm} engine to the parsnip \code{boost_tree} model
specification}
\usage{
add_boost_tree_lightgbm()
}
\description{
Wrapper to add \code{lightgbm} engine to the parsnip \code{boost_tree} model
specification
}
