% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{add_decision_tree_tree}
\alias{add_decision_tree_tree}
\title{Wrapper to add \code{tree} engine to the parsnip \code{decision_tree} model
specification}
\usage{
add_decision_tree_tree()
}
\description{
Wrapper to add \code{tree} engine to the parsnip \code{decision_tree} model
specification
}
