% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{multi_predict._lgb.Booster}
\alias{multi_predict._lgb.Booster}
\title{Model predictions across many sub-models}
\usage{
\method{multi_predict}{`_lgb.Booster`}(object, new_data, type = NULL, trees = NULL, ...)
}
\arguments{
\item{object}{A model_fit object.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or NULL. Possible values are "numeric", "class", "prob", "conf_int", "pred_int", "quantile", or "raw". When NULL, predict() will choose an appropriate value based on the model's mode.}

\item{trees}{An integer vector for the number of trees in the ensemble.}

\item{...}{Optional arguments to pass to predict.model_fit(type = "raw") such as type.}
}
\description{
For some models, predictions can be made on sub-models in the model object.
}
