% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{predict_lightgbm_classification_class}
\alias{predict_lightgbm_classification_class}
\title{predict_lightgbm_classification_class}
\usage{
predict_lightgbm_classification_class(object, new_data)
}
\arguments{
\item{object}{a fitted object.}

\item{new_data}{data frame in which to look for variables with which to predict.}
}
\description{
Not intended for direct use.
}
