% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{predict_lightgbm_regression_numeric}
\alias{predict_lightgbm_regression_numeric}
\title{predict_lightgbm_regression_numeric}
\usage{
predict_lightgbm_regression_numeric(object, new_data, ...)
}
\arguments{
\item{object}{a fitted object.}

\item{new_data}{data frame in which to look for variables with which to predict.}

\item{...}{Additional named arguments passed to the predict() method of the lgb.Booster object passed to object.}
}
\description{
Not intended for direct use.
}
