% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{train_catboost}
\alias{train_catboost}
\title{Boosted trees via catboost}
\usage{
train_catboost(
  x,
  y,
  depth = 6,
  iterations = 1000,
  learning_rate = NULL,
  rsm = 1,
  min_data_in_leaf = 1,
  subsample = 1,
  categorical_cols = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame or matrix of predictors}

\item{y}{A vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{depth}{An integer for the maximum depth of the tree.}

\item{iterations}{An integer for the number of boosting iterations.}

\item{learning_rate}{A numeric value between zero and one to control the learning rate.}

\item{rsm}{Subsampling proportion of columns.}

\item{min_data_in_leaf}{A numeric value for the minimum sum of instances needed
in a child to continue to split.}

\item{subsample}{Subsampling proportion of rows.}

\item{categorical_cols}{indices of categorical columns, when NULL (default) factor columns are automatically detected}

\item{...}{Other options to pass to \code{catboost.train}.}
}
\value{
A fitted \code{catboost.Model} object.
}
\description{
\code{catboost_train} is a wrapper for \code{catboost} tree-based models
where all of the model arguments are in the main function.
}
\keyword{internal}
