% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{train_lightgbm}
\alias{train_lightgbm}
\title{Boosted trees via lightgbm}
\usage{
train_lightgbm(
  x,
  y,
  max_depth = 17,
  num_iterations = 10,
  learning_rate = 0.1,
  feature_fraction = 1,
  min_data_in_leaf = 20,
  min_gain_to_split = 0,
  bagging_fraction = 1,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame or matrix of predictors}

\item{y}{A vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{max_depth}{An integer for the maximum depth of the tree.}

\item{num_iterations}{An integer for the number of boosting iterations.}

\item{learning_rate}{A numeric value between zero and one to control the learning rate.}

\item{feature_fraction}{Subsampling proportion of columns.}

\item{min_data_in_leaf}{A numeric value for the minimum sum of instances needed
in a child to continue to split.}

\item{min_gain_to_split}{A number for the minimum loss reduction required to make a
further partition on a leaf node of the tree.}

\item{bagging_fraction}{Subsampling proportion of rows.}

\item{quiet}{A logical; should logging by \code{\link[lightgbm:lgb.train]{lightgbm::lgb.train()}} be muted?}

\item{...}{Other options to pass to \code{\link[lightgbm:lgb.train]{lightgbm::lgb.train()}}.}
}
\value{
A fitted \code{lightgbm.Model} object.
}
\description{
\code{train_lightgbm} is a wrapper for \code{lightgbm} tree-based models
where all of the model arguments are in the main function.
}
\keyword{internal}
