% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{train_tree}
\alias{train_tree}
\title{Simple wrapper around \code{tree::tree}}
\usage{
train_tree(...)
}
\arguments{
\item{...}{arguments passed directly to \verb{[tree::tree()]}.}
}
\description{
Simple wrapper around \code{tree::tree} that adds a simple
class and sets the \code{mincut} argument to 0.
}
