% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesnip-package.R
\docType{package}
\name{treesnip-package}
\alias{treesnip}
\alias{treesnip-package}
\title{treesnip: Parsnip Engines for Additional Tree-Based Models}
\description{
\code{treesnip} provides engines for the \code{parsnip} package.
censored.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/curso-r/treesnip}
}

}
\author{
\strong{Maintainer}: Daniel Falbel \email{dfalbel@curso-r.com}

Authors:
\itemize{
  \item Athos Damiani \email{adamiani@curso-r.com}
  \item Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (\href{https://orcid.org/0000-0001-7509-0328}{ORCID})
}

}
\keyword{internal}
